+++ packages/code-du-travail-frontend/src/outils/IndemniteLicenciement/agreements/formula/index.ts
@@ -12,6 +12,8 @@ import { AgreementFormula2264 } from "./2264";
 import { AgreementFormula413 } from "./413";
 import { AgreementFormula650 } from "./650";
 import { AgreementFormula16 } from "./16";
+import { AgreementFormula573 } from "./573";
+import { AgreementFormula44 } from "./44";
 
 const getAgreementFormula = (
   idcc: SupportedCcIndemniteLicenciement,
@@ -52,6 +54,18 @@ const getAgreementFormula = (
       );
     case SupportedCcIndemniteLicenciement.IDCC0016 === idcc:
       return new AgreementFormula16().computeFormula(
+        agreementSeniority,
+        agreementRefSalary,
+        get
+      );
+    case SupportedCcIndemniteLicenciement.IDCC0573 === idcc:
+      return new AgreementFormula573().computeFormula(
+        agreementSeniority,
+        agreementRefSalary,
+        get
+      );
+    case SupportedCcIndemniteLicenciement.IDCC0044 === idcc:
+      return new AgreementFormula44().computeFormula(
         agreementSeniority,
         agreementRefSalary,
         get