+++ packages/code-du-travail-frontend/src/outils/IndemniteLicenciement/steps/Anciennete/store/__tests__/store.test.ts
@@ -73,6 +73,7 @@ describe("Ancienneté store", () => {
           key: MotifKeys.maladieNonPro,
           value: 1,
         },
+        startedAt: undefined,
       },
     ]);
     expect(store.getState().ancienneteData.input.dateEntree).toBe("20/02/2020");