+++ packages/code-du-travail-modeles/src/plugins/formule/index.ts
@@ -1,6 +1,8 @@
 import { SupportedCcIndemniteLicenciement } from "../types";
 import { Formula16 } from "./16_transports_routiers";
+import { Formula44 } from "./44_industries_chimiques";
 import { Formula413 } from "./413_handicap";
+import { Formula573 } from "./573_commerces_de_gros";
 import { Formula650 } from "./650_metallurgie_ingenieurs_cadres";
 import { Formula1516 } from "./1516_organismes_formation";
 import { Formula1596 } from "./1596_batiment_employes_ouvriers";
@@ -36,6 +38,10 @@ export class FormuleFactory {
         return new Formula650() as IFormula<T>;
       case SupportedCcIndemniteLicenciement.IDCC0016:
         return new Formula16() as IFormula<T>;
+      case SupportedCcIndemniteLicenciement.IDCC0573:
+        return new Formula573() as IFormula<T>;
+      case SupportedCcIndemniteLicenciement.IDCC0044:
+        return new Formula44() as IFormula<T>;
       case SupportedCcIndemniteLicenciement.default:
       default:
         return undefined;