+++ packages/code-du-travail-modeles/src/plugins/formule/types.ts
@@ -1,6 +1,8 @@
 import type { SupportedCcIndemniteLicenciement } from "..";
 import type { Formula16Props } from "./16_transports_routiers";
+import type { Formula44Props } from "./44_industries_chimiques";
 import type { Formula413Props } from "./413_handicap";
+import type { Formula573Props } from "./573_commerces_de_gros";
 import type { Formula650Props } from "./650_metallurgie_ingenieurs_cadres";
 import type { Formula1596Props } from "./1596_batiment_employes_ouvriers";
 import type { Formula1597Props } from "./1597_batiment_employes_ouvriers_bis";
@@ -39,4 +41,8 @@ export type FormulaProps<T> =
     ? Formula650Props
     : T extends SupportedCcIndemniteLicenciement.IDCC0016
     ? Formula16Props
+    : T extends SupportedCcIndemniteLicenciement.IDCC0573
+    ? Formula573Props
+    : T extends SupportedCcIndemniteLicenciement.IDCC0044
+    ? Formula44Props
     : DefaultFormulaProps;